factory method createRenderingEngine; moved content-related and non-OGL methods to src/core/contents/
